feat: Finished split codes
